fixup! Refactor TransactionContext to serialise session-related information
