+++ benchmarks/src/test/java/io/crate/execution/engine/reader/JsonReaderBenchmark.java
@@ -1,7 +1,7 @@
 package io.crate.execution.engine.reader;
 
 import com.google.common.collect.ImmutableMap;
-import io.crate.action.sql.SessionTransportableInfo;
+import io.crate.metadata.settings.SessionTransportableInfo;
 import io.crate.data.BatchIterator;
 import io.crate.data.Input;
 import io.crate.data.Row;
