+++ sql/src/test/java/io/crate/execution/dml/upsert/TransportShardUpsertActionTest.java
@@ -23,7 +23,7 @@
 package io.crate.execution.dml.upsert;
 
 import io.crate.Constants;
-import io.crate.action.sql.SessionTransportableInfo;
+import io.crate.metadata.settings.SessionTransportableInfo;
 import io.crate.exceptions.InvalidColumnNameException;
 import io.crate.execution.ddl.SchemaUpdateClient;
 import io.crate.execution.dml.ShardResponse;
